backend/requirements.txt
bleach==6.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
Deprecated==1.3.1
Flask==3.0.3
flask-cors==6.0.2
Flask-Limiter==3.7.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.19.2
python-dotenv==1.0.1
python-http-client==3.3.7
rich==13.9.4
sendgrid==6.12.5
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.6
wrapt==2.1.1
WTForms==3.2.1
